requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
executing==1.2.0
fonttools==4.42.1
frozenlist==1.4.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.25.2
openai==0.27.9
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
yarl==1.9.2
